Moved validation middleware in its own file
